Add Brian and Rodney
Rodney is showing a RAL affiliation since that is where he did
the work. I wonder if the affiliation should be something instead
along the lines of

 "formerly of CLRC"

Ordinarily you put the affiliation where the work was done and
a footnote to present affiliation but that's not really relevant
in this case.