Fix a couple of typos.
Some years ago I read in the Chicago Manual of style that the word
after a hyphen in a compound noun does not begin with a capital.  So
it would N-dimensional Data Format.  It depends whether you think the
D in NDF referes to Dimensional or Data.  Anyway I've just made it
consistent and the way SGP/38 does it.