Refactored TABLE section.
Since the original first paragraph was written by Tim without
knowledge of the CHI/tbl format, and the original note requested a
writeup of the tabular format created bt FITS2NDF, that's what I did.
This is an attempt to meld the two parts in a more-cohesive way.
It probably needs another iteration.

The actual extended TABLE format made by FITS2NDF is now described and
illustrated in an appendix.