Use shorter forms of journal names
A&C may not like it but for now it's what I need to do to get the
PDF to compile. The problem is that the George 1993 reference splits
the URL over page 15 and 16 and hypernat does not like that. Shortening
the journal names allows it to fit entirely on page 15.